eth_getBlock properly returns transactions
